+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `Config::{inbound_queue_depth, initial_read_buffer_size}` were moved to `Reading::{MESSAGE_QUEUE_DEPTH, INITIAL_BUFFER_SIZE}`
 - `Config::max_handshake_time_ms` was moved to `Handshake::TIMEOUT_MS`
 - `Config::outbound_queue_depth` was moved to `Writing::MESSAGE_QUEUE_DEPTH`
+- `Writing::send_direct_message` now returns an `io::Result<oneshot::Receiver<io::Result<()>>>` instead of `io::Result<oneshot::Receiver<bool>>`
 
 # 0.36.0